+++ src/ui/setup/pages/SocialFlow/types/auto-imports.d.ts
@@ -266,6 +266,7 @@ declare global {
   const useScrollLock: typeof import('@vueuse/core')['useScrollLock']
   const useSessionStorage: typeof import('@vueuse/core')['useSessionStorage']
   const useShare: typeof import('@vueuse/core')['useShare']
+  const useSignupNudge: typeof import('../../../../../composables/useSignupNudge')['useSignupNudge']
   const useSlots: typeof import('vue')['useSlots']
   const useSocialNetworksStore: typeof import('../../../../../stores/socialNetworks')['useSocialNetworksStore']
   const useSorted: typeof import('@vueuse/core')['useSorted']
@@ -346,9 +347,6 @@ declare global {
   export type { KanbanItem, KanbanColumnId } from '../../../../../stores/kanban'
   import('../../../../../stores/kanban')
   // @ts-ignore
-  export type { FacebookPost } from '../../../../../stores/mockData/facebookMock'
-  import('../../../../../stores/mockData/facebookMock')
-  // @ts-ignore
   export type { Profile } from '../../../../../stores/profiles'
   import('../../../../../stores/profiles')
 }
@@ -361,7 +359,6 @@ declare module 'vue' {
     readonly EffectScope: UnwrapRef<typeof import('vue')['EffectScope']>
     readonly WEBVIEW_URLS: UnwrapRef<typeof import('../../../../../stores/webviewState')['WEBVIEW_URLS']>
     readonly acceptHMRUpdate: UnwrapRef<typeof import('pinia')['acceptHMRUpdate']>
-    readonly appRouter: UnwrapRef<typeof import('../../../../../utils/router/index')['appRouter']>
     readonly applyDisableCopyProtection: UnwrapRef<typeof import('../../../../../utils/disableCopyProtection')['applyDisableCopyProtection']>
     readonly asyncComputed: UnwrapRef<typeof import('@vueuse/core')['asyncComputed']>
     readonly autoResetRef: UnwrapRef<typeof import('@vueuse/core')['autoResetRef']>
@@ -412,7 +409,6 @@ declare module 'vue' {
     readonly mapStores: UnwrapRef<typeof import('pinia')['mapStores']>
     readonly mapWritableState: UnwrapRef<typeof import('pinia')['mapWritableState']>
     readonly markRaw: UnwrapRef<typeof import('vue')['markRaw']>
-    readonly mockPosts: UnwrapRef<typeof import('../../../../../stores/mockData/facebookMock')['mockPosts']>
     readonly nextTick: UnwrapRef<typeof import('vue')['nextTick']>
     readonly notivue: UnwrapRef<typeof import('../../../../../utils/notifications')['notivue']>
     readonly onActivated: UnwrapRef<typeof import('vue')['onActivated']>
@@ -545,7 +541,6 @@ declare module 'vue' {
     readonly useEventListener: UnwrapRef<typeof import('@vueuse/core')['useEventListener']>
     readonly useEventSource: UnwrapRef<typeof import('@vueuse/core')['useEventSource']>
     readonly useEyeDropper: UnwrapRef<typeof import('@vueuse/core')['useEyeDropper']>
-    readonly useFacebookMockStore: UnwrapRef<typeof import('../../../../../stores/mockData/facebookMock')['useFacebookMockStore']>
     readonly useFavicon: UnwrapRef<typeof import('@vueuse/core')['useFavicon']>
     readonly useFetch: UnwrapRef<typeof import('@vueuse/core')['useFetch']>
     readonly useFileDialog: UnwrapRef<typeof import('@vueuse/core')['useFileDialog']>
@@ -618,6 +613,7 @@ declare module 'vue' {
     readonly useScrollLock: UnwrapRef<typeof import('@vueuse/core')['useScrollLock']>
     readonly useSessionStorage: UnwrapRef<typeof import('@vueuse/core')['useSessionStorage']>
     readonly useShare: UnwrapRef<typeof import('@vueuse/core')['useShare']>
+    readonly useSignupNudge: UnwrapRef<typeof import('../../../../../composables/useSignupNudge')['useSignupNudge']>
     readonly useSlots: UnwrapRef<typeof import('vue')['useSlots']>
     readonly useSocialNetworksStore: UnwrapRef<typeof import('../../../../../stores/socialNetworks')['useSocialNetworksStore']>
     readonly useSorted: UnwrapRef<typeof import('@vueuse/core')['useSorted']>